refactor(operator_commands/dispatch): merge inline tests into tests_dispatch.rs (#1266)
Inline #[cfg(test)] mod tests block in dispatch.rs is moved into the
existing sibling tests_dispatch.rs as a nested submodule 'dispatch_inline'
to scope its 'fn args' helper (which collides with the existing helper of
the same name in tests_dispatch.rs).

- src/operator_commands/dispatch.rs: 413 → 346 LOC (under 400-line rule)
- src/operator_commands/tests_dispatch.rs: 386 → 453 LOC

Pure refactor: zero logic changes, all 518 operator_commands tests pass.

Closes 1/70 of #1266 (after #1285-#1290).
