+++ src/operator_commands/dispatch.rs
@@ -343,71 +343,3 @@ pub fn dispatch_probe_with_context(
 
     Ok(())
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    fn args(items: &[&str]) -> Vec<String> {
-        items.iter().map(|s| s.to_string()).collect()
-    }
-
-    #[test]
-    fn next_required_returns_value() {
-        let mut it = args(&["hello", "world"]).into_iter();
-        assert_eq!(next_required(&mut it, "first").unwrap(), "hello");
-        assert_eq!(next_required(&mut it, "second").unwrap(), "world");
-    }
-
-    #[test]
-    fn next_required_error_on_empty() {
-        let mut it = std::iter::empty::<String>();
-        assert!(next_required(&mut it, "missing").is_err());
-    }
-
-    #[test]
-    fn next_optional_path_some_and_none() {
-        let mut it = args(&["/tmp/test"]).into_iter();
-        let p = next_optional_path(&mut it);
-        assert_eq!(p, Some(PathBuf::from("/tmp/test")));
-
-        let mut it = std::iter::empty::<String>();
-        assert_eq!(next_optional_path(&mut it), None);
-    }
-
-    #[test]
-    fn reject_extra_args_ok_when_empty() {
-        assert!(reject_extra_args(std::iter::empty::<String>()).is_ok());
-    }
-
-    #[test]
-    fn reject_extra_args_err_with_extra() {
-        let result = reject_extra_args(args(&["extra1", "extra2"]).into_iter());
-        assert!(result.is_err());
-        let msg = format!("{}", result.unwrap_err());
-        assert!(msg.contains("extra1"));
-        assert!(msg.contains("extra2"));
-    }
-
-    #[test]
-    fn gym_usage_returns_static_str() {
-        let usage = gym_usage();
-        assert!(usage.contains("simard-gym"));
-        assert!(usage.contains("list"));
-        assert!(usage.contains("run-suite"));
-    }
-
-    #[test]
-    fn dispatch_operator_probe_unknown_command() {
-        let result = dispatch_operator_probe(vec!["nonexistent-command".to_string()]);
-        assert!(result.is_err());
-        let msg = format!("{}", result.unwrap_err());
-        assert!(msg.contains("unsupported"));
-    }
-
-    #[test]
-    fn dispatch_legacy_gym_cli_no_args() {
-        let result = dispatch_legacy_gym_cli(std::iter::empty::<String>());
-        assert!(result.is_err());
-    }
-}
